+++ src/main/kotlin/com/coder/toolbox/views/CoderPage.kt
@@ -6,8 +6,6 @@ import com.jetbrains.toolbox.api.core.ui.icons.SvgIcon.IconType
 import com.jetbrains.toolbox.api.localization.LocalizableString
 import com.jetbrains.toolbox.api.ui.actions.RunnableActionDescription
 import com.jetbrains.toolbox.api.ui.components.UiPage
-import kotlinx.coroutines.launch
-import java.util.UUID
 
 /**
  * Base page that handles the icon, displaying error notifications, and
@@ -37,21 +35,6 @@ abstract class CoderPage(
         )
     } else {
         SvgIcon(byteArrayOf(), type = IconType.Masked)
-    }
-
-    /**
-     * Show an error as a popup on this page.
-     */
-    fun notify(logPrefix: String, ex: Throwable) {
-        context.logger.error(ex, logPrefix)
-        context.cs.launch {
-            context.ui.showSnackbar(
-                UUID.randomUUID().toString(),
-                context.i18n.pnotr(logPrefix),
-                context.i18n.pnotr(ex.message ?: ""),
-                context.i18n.ptrl("Dismiss")
-            )
-        }
     }
 
     companion object {